refactor package structure for AddStateToProjectUseCase and its test
